requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
gradio>=4.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0,<2.0.0
Pillow>=9.5.0
seaborn>=0.12.0
huggingface_hub>=0.16.0